Continued adding scene system.
All children of entities are now added when you add them to the scene.
Engine can now have the scene set.

Added the RenderPlanner which takes a scene and works out how to render
it. Currently it is just setting up the uniform buffers.
Added PlannerData component to hold internal values.

Added a debug UI for the scene which allows you to see and modify scene
parameters.