Fix Pipeline not reserving enough descriptors when mixing PerSwapChain and Single bindings
If you were to use a Camera binding and also use some other plain binding, it would not reserve enough bindings for all the plain bindings.
This is because it needs X sets of each binding, where X is the higher number of each binding. So the camera needs 3 but the plain one needed 1.
Previously it would only reserve 1 of the plain. Now it also reserves 3.