Started building out shader system
The shader system will allow for rendering Entities with specific shaders.
You will be able to override specific stages (vertex, fragment, etc), or all the stages.
Eventually this will be used with Material. It will know which Shader is attached, and the options it has (textures and other settings) will be dependant on the Shader.

This commit introduces the various shader classes and builders.
It also introduces the standard built-in shaders as ShaderStages which are loaded when the engine starts up.

The system also automatically checks compatibility with the other stages and pipelines.