Started adding Shader support to RenderPlanner
Entities to render are now grouped by Shader, since every Shader requires its own pipeline.
When you add an Entity to be rendered, it adds the Entity to the relevant group, creating it if needed.
If it creates a group it also instructs the DeferredPipeline to prepare a pipeline for that shader.

When removing an Entity from render, it does the opposite. If there are no Entities being rendered with a previously used Shader, then it will instruct DeferredPipeline to remove the pipeline.

This also adds the ability to update any of the render parameters.