Implemented material uniform variables
These are strongly typed but generic. ShaderUniformId objects are used to specify the variable you want to modify.

The MaterialDescriptorCache is now able to create and manage Buffers to hold the variable information.

Currently the buffers are not optimal. Each uniform gets its own buffer and they are all Host Visible buffers.

Eventually it would be nice to provide a single buffer which holds all uniforms for the Material, and then make that DeviceLocal only.